+++ scripts/download-nodeSqlite3Wasm-docs.js
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Universal Doc Downloader & Cleaner Template
  * -------------------------------------------